Restore tests module to its original place as well.
