+++ packages/yew/src/lib.rs
@@ -286,9 +286,7 @@ mod renderer;
 
 #[cfg(feature = "render")]
 #[cfg(test)]
-pub mod tests {
-    pub use crate::dom_bundle::layout_tests;
-}
+pub mod tests;
 
 /// The module that contains all events available in the framework.
 pub mod events {
