+++ packages/yew/src/tests/mod.rs
@@ -1,0 +1,1 @@
+pub mod layout_tests;
